Remove affirmative upper/lower/digit flags
The default is to include classes, so having flags to include them
doesn't do anything useful. To fix this, refactor the code so that the
flags are now to exclude rather than include classes, retaining the
same `--no-upper/--no-lower/--no-digit` names but making them flags in
their own right instead of inversions of `--upper/--lower/--digit`.